animal_posts_service/requirements.txt
grpcio
grpcio-tools
Flask
requests
SQLAlchemy
Flask-SQLAlchemy
prometheus-flask-exporter
